backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
reportlab==4.0.8
nltk==3.8.1
wcwidth==0.2.13
python-docx==1.0.1
pypdf==3.17.4
Pillow==10.0.1
